server/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
dataclasses-json==0.6.3
Flask==3.0.1
Flask-Cors==4.0.0
frozenlist==1.4.1
gigachain==0.1.0.1
gigachain-community==0.0.11.1
gigachain-core==0.1.9.1
gigachat==0.1.14
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langsmith==0.0.83
MarkupSafe==2.1.4
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
protobuf==4.25.2
pycparser==2.21
pydantic==1.10.14
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
yandexcloud==0.257.0
yarl==1.9.4
